Reworked order and Update
Made Order function look cleaner while also making it so the number of swipes will now be updated. Next step should be fixing the issue with the order function taking away a swipe no matter what the cost of the item itself is. Text me if you have any questions.